requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=6.2.5
python-multipart>=0.0.5
python-dotenv>=0.19.0
requests
pydantic_settings